rqe_iterators: factor out new_missing_iterator()
The QAST eval code will need the exact same code.
