+++ src/redisearch_rs/c_entrypoint/iterators_ffi/src/inverted_index/missing.rs
@@ -7,105 +7,10 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 
-use std::{fmt::Debug, ptr::NonNull};
+use std::ptr::NonNull;
 
-use field::{FieldExpirationPredicate, FieldFilterContext, FieldMaskOrIndex};
-use index_result::RSIndexResult;
-use inverted_index::{IndexReader, doc_ids_only::DocIdsOnly, raw_doc_ids_only::RawDocIdsOnly};
-use rqe_core::{DocId, FieldIndex};
-use rqe_iterators::{
-    FieldExpirationChecker, IteratorType, interop::RQEIteratorWrapper, inverted_index::Missing,
-    profile_print,
-};
-
-/// Wrapper around different II missing iterator encoding types to avoid generics in FFI code.
-///
-/// Handles both the standard variable-length encoding ([`DocIdsOnly`]) and the
-/// fixed 4-byte raw encoding ([`RawDocIdsOnly`]).
-pub(super) enum MissingIterator<'index> {
-    Encoded(Missing<'index, DocIdsOnly, FieldExpirationChecker>),
-    Raw(Missing<'index, RawDocIdsOnly, FieldExpirationChecker>),
-}
-
-impl Debug for MissingIterator<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let variant = match self {
-            MissingIterator::Encoded(_) => "Encoded",
-            MissingIterator::Raw(_) => "Raw",
-        };
-        write!(f, "MissingIterator({variant})")
-    }
-}
-
-/// Macro to dispatch RQEIterator methods across all [`MissingIterator`] variants.
-macro_rules! dispatch {
-    ($self:expr, $method:ident $(, $arg:expr)*) => {
-        match $self {
-            MissingIterator::Encoded(m) => m.$method($($arg),*),
-            MissingIterator::Raw(m) => m.$method($($arg),*),
-        }
-    };
-}
-
-impl<'index> rqe_iterators::RQEIterator<'index> for MissingIterator<'index> {
-    #[inline(always)]
-    fn current(&mut self) -> Option<&mut RSIndexResult<'index>> {
-        dispatch!(self, current)
-    }
-
-    #[inline(always)]
-    fn read(
-        &mut self,
-    ) -> Result<Option<&mut RSIndexResult<'index>>, rqe_iterators::RQEIteratorError> {
-        dispatch!(self, read)
-    }
-
-    #[inline(always)]
-    fn skip_to(
-        &mut self,
-        doc_id: DocId,
-    ) -> Result<Option<rqe_iterators::SkipToOutcome<'_, 'index>>, rqe_iterators::RQEIteratorError>
-    {
-        dispatch!(self, skip_to, doc_id)
-    }
-
-    #[inline(always)]
-    fn rewind(&mut self) {
-        dispatch!(self, rewind)
-    }
-
-    #[inline(always)]
-    fn num_estimated(&self) -> usize {
-        dispatch!(self, num_estimated)
-    }
-
-    #[inline(always)]
-    fn last_doc_id(&self) -> DocId {
-        dispatch!(self, last_doc_id)
-    }
-
-    #[inline(always)]
-    fn at_eof(&self) -> bool {
-        dispatch!(self, at_eof)
-    }
-
-    #[inline(always)]
-    fn revalidate(
-        &mut self,
-        spec: &index_spec::IndexSpecReadGuard,
-    ) -> Result<rqe_iterators::RQEValidateStatus<'_, 'index>, rqe_iterators::RQEIteratorError> {
-        dispatch!(self, revalidate, spec)
-    }
-
-    #[inline(always)]
-    fn type_(&self) -> IteratorType {
-        IteratorType::InvIdxMissing
-    }
-
-    fn intersection_sort_weight(&self, _prioritize_union_children: bool) -> f64 {
-        1.0
-    }
-}
+use rqe_core::FieldIndex;
+use rqe_iterators::{interop::RQEIteratorWrapper, inverted_index::new_missing_iterator};
 
 /// Creates a new missing-field inverted index iterator.
 ///
@@ -140,56 +45,15 @@ pub unsafe extern "C" fn NewInvIndIterator_MissingQuery(
     debug_assert!(!idx.is_null(), "idx must not be null");
     debug_assert!(!sctx.is_null(), "sctx must not be null");
 
-    // Cast to the FFI wrapper enum which handles type dispatch
     let idx_ffi: *const inverted_index_ffi::InvertedIndex = idx.cast();
-    // SAFETY: 1. guarantees idx is valid and non-null
+    // SAFETY: 1. guarantees idx is valid and non-null.
     let ii_ref = unsafe { &*idx_ffi };
 
-    // SAFETY: 3. guarantees sctx is valid and non-null
+    // SAFETY: 3. guarantees sctx is valid and non-null.
     let sctx_nn = unsafe { NonNull::new_unchecked(sctx as *mut _) };
 
-    // Build the expiration checker for the Missing predicate
-    let filter_ctx = FieldFilterContext {
-        field: FieldMaskOrIndex::Index(field_index),
-        predicate: FieldExpirationPredicate::Missing,
-    };
-
-    // Create the appropriate missing iterator variant based on encoding type.
-    let iterator = match ii_ref {
-        inverted_index_ffi::InvertedIndex::DocIdsOnly(ii) => {
-            let reader = ii.reader();
-            // SAFETY: 3.-6. guarantee validity for the iterator's lifetime.
-            let checker =
-                unsafe { FieldExpirationChecker::new(sctx_nn, filter_ctx, reader.flags()) };
-            // SAFETY: 3.-6. guarantee validity for the iterator's lifetime.
-            MissingIterator::Encoded(unsafe { Missing::new(reader, sctx_nn, field_index, checker) })
-        }
-        inverted_index_ffi::InvertedIndex::RawDocIdsOnly(ii) => {
-            let reader = ii.reader();
-            // SAFETY: 3.-6. guarantee validity for the iterator's lifetime.
-            let checker =
-                unsafe { FieldExpirationChecker::new(sctx_nn, filter_ctx, reader.flags()) };
-            // SAFETY: 3.-6. guarantee validity for the iterator's lifetime.
-            MissingIterator::Raw(unsafe { Missing::new(reader, sctx_nn, field_index, checker) })
-        }
-        _ => panic!(
-            "Missing iterator requires a DocIdsOnly or RawDocIdsOnly inverted index, got: {:?}",
-            std::mem::discriminant(ii_ref)
-        ),
-    };
-
-    RQEIteratorWrapper::boxed_new(iterator)
-}
-
-impl profile_print::ProfilePrint for MissingIterator<'_> {
-    fn print_profile(
-        &self,
-        map: &mut redis_reply::MapBuilder<'_>,
-        ctx: &mut profile_print::ProfilePrintCtx<'_>,
-    ) {
-        match self {
-            MissingIterator::Encoded(m) => m.print_profile(map, ctx),
-            MissingIterator::Raw(m) => m.print_profile(map, ctx),
-        }
-    }
+    // SAFETY: 3.-6. guarantee sctx, spec, field_index, and
+    // missingFieldDict validity.
+    let it = unsafe { new_missing_iterator(ii_ref, sctx_nn, field_index) };
+    RQEIteratorWrapper::boxed_new(it)
 }
