+++ src/redisearch_rs/rqe_iterators/src/inverted_index/mod.rs
@@ -22,7 +22,7 @@ pub use geo::{
     GeoRangeError, InvalidGeoInput, build_geo_numeric_filters, extract_geo_unit_factor,
     new_geo_range_iterator,
 };
-pub use missing::Missing;
+pub use missing::{Missing, new_missing_iterator};
 pub use numeric::{Numeric, NumericIteratorVariant, open_numeric_or_geo_index};
 pub use tag::Tag;
 pub use term::Term;
